made usersReducer module
